move most auth helpers from shared/middleware to dashboard_legacy
