+++ dashboard_legacy/test/middleware/helpers/test_auth_helpers.rb
@@ -1,4 +1,4 @@
-require_relative '../../test_helper'
+require_relative '../middleware_test_helper'
 require_relative '../../../middleware/helpers/auth_helpers'
 
 class AuthHelpersTest < Minitest::Test
